+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # master
 
+# 5.0.0
+  * fix asset-url transformation for lazy engines with useDeprecatedIncorrectCssProcesing flag
+  * assetHttpPrefix for lazy engines now always includes engines-dist/<engine.name>
+
 # 4.0.0